adds a fakeApplicationBuilder trait
