+++ test/uk/gov/hmrc/cataloguefrontend/FrontendRouteWarningsPageSpec.scala
@@ -17,20 +17,10 @@
 package uk.gov.hmrc.cataloguefrontend
 
 import com.github.tomakehurst.wiremock.http.RequestMethod._
-import org.scalatestplus.play.guice.GuiceOneServerPerSuite
-import play.api.Application
-import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.ws._
 import uk.gov.hmrc.cataloguefrontend.util.UnitSpec
 
-class FrontendRouteWarningsPageSpec extends UnitSpec with GuiceOneServerPerSuite with WireMockEndpoints {
-
-  override def fakeApplication: Application = new GuiceApplicationBuilder()
-    .configure(
-      "microservice.services.shutter-api.port"          -> endpointPort,
-      "microservice.services.shutter-api.host"          -> host
-    )
-    .build()
+class FrontendRouteWarningsPageSpec extends UnitSpec with FakeApplicationBuilder {
 
   private[this] lazy val ws = app.injector.instanceOf[WSClient]
