+++ test/uk/gov/hmrc/cataloguefrontend/TeamsSpec.scala
@@ -16,28 +16,14 @@
 
 package uk.gov.hmrc.cataloguefrontend
 
-import java.time.{LocalDateTime, ZoneOffset}
-
 import com.github.tomakehurst.wiremock.http.RequestMethod._
 import org.scalatest.BeforeAndAfter
-import org.scalatestplus.play.guice.GuiceOneServerPerSuite
-import play.api.Application
-import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.ws._
 import uk.gov.hmrc.cataloguefrontend.util.UnitSpec
 
-class TeamsSpec extends UnitSpec with BeforeAndAfter with GuiceOneServerPerSuite with WireMockEndpoints {
+import java.time.{LocalDateTime, ZoneOffset}
 
-  override def fakeApplication: Application =
-    new GuiceApplicationBuilder()
-      .configure(Map(
-        "microservice.services.teams-and-repositories.port" -> endpointPort,
-        "microservice.services.teams-and-repositories.host" -> host,
-        "play.ws.ssl.loose.acceptAnyCertificate"            -> true,
-        "play.http.requestHandler"                          -> "play.api.http.DefaultHttpRequestHandler",
-        "metrics.jvm"                                       -> false
-      ))
-      .build()
+class TeamsSpec extends UnitSpec with BeforeAndAfter with FakeApplicationBuilder {
 
   private[this] lazy val WS = app.injector.instanceOf[WSClient]
